+++ src/main/resources/com/github/jlangch/venice/repl.sh
@@ -6,7 +6,7 @@
 #    |
 #    +-- libs
 #    |    +-- repl.json
-#    |    +-- venice-1.10.29.jar
+#    |    +-- venice-1.10.30.jar
 #    |
 #    +-- scripts
 #    |    +-- script-1.venice